user-service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
pika
pillow
django-cors-headers
django-tenants
python-dotenv
requests
python-dotenv
faker